Update to 4.25
Updated to UE4 4.25. Updated the inputs so that it is ready for OpenXR plugin support. Added a C++ BlueprintFunctionLibrary to switch the hands inputs because the Oculus Controller keys are only in C++ now. The project is correctly named as First person (FP) project. Only works with Oculus Right now until OpenXR plugin works with the runtime working and the other inputs can be emulated correctly.